Add missing flag initialization
The delay buffer `stopped` field was not initialized.

Since it practice the unique instance of sc_delay_buffer is initialized
in static memory, the flag was initialized to false as a side effect.

But with commit aec64633eac571f4c131fe42e37bf35f0e0aac31, in debug mode
only, the delay buffer was broken.