Avoid resetting pending frame
The function update_texture() calls update_texture_internal() and falls
back to set_pending_frame() if it fails.

When the frame passed is the pending frame, call only the _internal()
version instead.

This will prevent issues with frame reference counts by ensuring the
source and destination frames are never the same.

Refs 6d24175f4ec34aeb38184df2aa287fe9b5b9f39a
Refs #6357 <https://github.com/Genymobile/scrcpy/issues/6357>